# aufs optimizations