1. Remove global Get()/Put() locking, add lock around mount().
2. Streamline Put() (unmount).
3. Optimize lots-of-layers case.
3. Simplify Cleanup().

For detailed description, see commit messages.

### Performance and stress testing

This was tested by the following script on a machine with 8GB RAM (with more RAM you can increase the value of `PARALLEL`):

```bash
#!/bin/bash

NUMCTS=1000
PARALLEL=100
IMAGE=golang

create() {
	# -v 'vol_{}:/vol'
	time ( seq $NUMCTS | parallel -j$PARALLEL docker create $IMAGE top > /dev/null )
	echo "^^^ CREATE TIME ^^^"
}

remove() {
	time ( docker ps -qa | shuf | tail -n $NUMCTS | parallel -j$PARALLEL docker rm -f '{}' > /dev/null )
	echo "^^^ REMOVE TIME ^^^"
}

## MAIN

docker pull $IMAGE

# stage 1: precreate $NUMCTS containers
create
echo

# stage2: create another $NUMCTS while removing $NUMCTS containers
create &
remove &
wait

# stage 3: remove the rest of it
remove
```

Results for the above tests, measured at least 3 times (with https://github.com/moby/moby/pull/39135 applied, with and without this PR):
 - container creation on stage 1 is 3x faster (from 60s to 20s);
 - container creation while removal (stage 2) is almost 2x faster (from 85s to 45s);
 - container removal (stages 2 and 3) times are about the same;

No errors were detected.