@thaJeztah 2019-04-18
```
03:02:27 daemon/graphdriver/aufs/aufs.go:604:11:warning: undeclared name: x (gosimple)
03:02:27 daemon/graphdriver/aufs/aufs.go:426:11:warning: a.unmount undefined (type *Driver has no field or method unmount) (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:566:21:warning: cannot use m (variable of type os.FileInfo) as string value in argument to Unmount (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:604:11:warning: undeclared name: x (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:314:6:warning: undeclared name: err (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:313:3:warning: undeclared name: err (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:313:11:warning: a.unmount undefined (type *Driver has no field or method unmount) (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:322:24:warning: undeclared name: err (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:319:24:warning: undeclared name: err (unconvert)
03:02:27 daemon/graphdriver/aufs/aufs.go:318:19:warning: undeclared name: err (unconvert)
```