@kolyshkin 2019-05-03
> I'm pretty sure I tried to remove this lock before and ran into trouble.

@cpuguy83 yeah, it took me quite a while (days) to figure out what the proper locking should be (and I have yet to update this PR with my new patchset). Apparently, parallel unmounts is fine, but there's a race in aufs kernel module that prevents parallel mounts, so this has to be done under a lock.

Will try to bring this back to life today.