@kolyshkin 2019-05-03
OK, I have updated the patch set and did some benchmarks (on top of #39135, using the script from there). Parallel container creation is now twice faster, other timings are similar.