@thaJeztah 2019-05-03
@kolyshkin build is failing;

```
18:54:56 # github.com/docker/docker/daemon/graphdriver/aufs
18:54:56 daemon/graphdriver/aufs/aufs.go:626: Sprintf format %s reads arg #1, but call has 0 args
18:54:57 FAIL	github.com/docker/docker/daemon/graphdriver/aufs [build failed]
```