@kolyshkin 2019-05-03
Not sure if this is relevant (error on exp CI):

> 00:14:27.831 --- FAIL: TestServiceCreateInit (5.81s)
> 00:14:27.831     --- FAIL: TestServiceCreateInit/daemonInitDisabled (2.45s)
> 00:14:27.831         create_test.go:38: Creating a new daemon
> 00:14:27.831         daemon.go:288: [de58392204bd7] waiting for daemon to start
> 00:14:27.831         daemon.go:320: [de58392204bd7] daemon started
> 00:14:27.831         create_test.go:52: assertion failed: error is not nil: Error response from daemon: rpc error: code = AlreadyExists desc = name conflicts with an existing object: service gracious_ramanujan already exists: error creating service
> 00:14:27.831         daemon.go:275: [de58392204bd7] exiting daemon