@thaJeztah 2019-05-03
perhaps the name generator wasn't random enough? Interesting; never encountered that