@kolyshkin 2019-05-03
janky CI is flaky test (https://github.com/moby/moby/issues/32673)
> 01:04:57.140 FAIL: docker_api_swarm_test.go:292: DockerSwarmSuite.TestAPISwarmLeaderElection