@kolyshkin 2019-05-04
Windows CI failure, def unrelated (as this PR changes aufs graph driver used by Linux only)
```
00:39:15.779 
00:39:15.779 ----------------------------------------------------------------------
00:39:15.779 FAIL: check_test.go:107: DockerSuite.TearDownTest
00:39:15.779 
00:39:15.779 assertion failed: error is not nil: Error response from daemon: container ddb0019e76d456354c4761d6426da149fe516d1eb437737e4a1fc26ef03f4b67: driver "windowsfilter" failed to remove root filesystem: hcsshim::GetComputeSystems: The requested compute system operation is not valid in the current state.: failed to remove ddb0019e76d456354c4761d6426da149fe516d1eb437737e4a1fc26ef03f4b67
00:39:15.779 
```