@kolyshkin 2019-05-04
I am seeing TestService* test timeouts after 5m when run as
```
export TEST_REPEAT=10
export TESTFLAGS="-test.count 10 -test.run TestService"
./hack/make.sh test-integration
```
even without the patches from this PR (see https://github.com/moby/moby/pull/39171), so this PR is not the one to blame.