@kolyshkin 2019-05-04
OK, I am really doing something stupid here. CI does not test aufs driver, so its results are irrelevant.