@kolyshkin 2019-05-06
> 01:28:27.996 FAIL: docker_cli_swarm_test.go:1317: DockerSwarmSuite.TestSwarmClusterRotateUnlockKey

is a known flaky test; this is the only CI failure