@kolyshkin 2019-05-07
> I would like to get a chance to give this a thorough manual test.

Yes please; I hope you can think of something else than the test I come up with (it's in the description of this thread). If you'll run mine, make sure to change PARALLEL value to suit your RAM to avoid ENOMEM -- in my case 100 is good for 8GB, 150 occasionally gives ENOMEM.

> I recall before seeing issues with the auplink call.

The issue I was seeing was a race in the aufs kernel module, something about xino -- so I ended up not allowing more than a single `mount -t aufs` at a time.

The only issue I saw with `auplink flush` is when it's called on an error path where the aufs mount failed, and it fails because there's no mount. But maybe my testing is just not good enough.