@kolyshkin 2019-05-07
NOTE: I just modified one of the commits (a0d0b88b5cc1729e "aufs: use mount.Unmount") so it no longer removes a `mounted` check from Remove() method.