@cpuguy83 2019-05-09
So there is a benchmark in aufs test dir for concurrent access.
WIth this change I see a bunch of:

```
WARN[0001] Couldn't run auplink before unmount /tmp/aufs-tests/aufs/mnt/3ed1ff0d4927f4741265c6eefb3859d9bccd47c919ec5b053a3c380fcd4da6ce  error="exit status 22" storage-driver=aufs
```