@dmcgowan 2019-05-09
I can see how the global lock is unnecessary, but the lock should at least be on the target, which this PR is not doing. I don't understand what putting a lock around `unix.Mount` is accomplishing since there is no additional check after the lock is acquired.