@kolyshkin 2019-05-09
@cpuguy83 thanks, I'll take a look (I also confirm that there are no such warnings before this patch).

> but the lock should at least be on the target, which this PR is not doing

@dmcgowan quoting my commit message
>    Both mount and unmount calls are already protected by fine-grained
>    (per id) locks in Get()/Put() introduced in commit fc1cf1911bb
>    ("Add more locking to storage drivers"), so there's no point in
>    having a global lock in mount/unmount.

So there is per-id (which can be though of a short for mount point) locking in place already. Let me know if you think it's not adequate, or say the lock should really be on a per mount-point rather than just id.

> I don't understand what putting a lock around unix.Mount is accomplishing
@dmcgowan as described in patch commit message, I have discovered that doing two mounts in parallel (with different mount points and arguments, i.e. unrelated ones) sometimes hit a bug in the aufs kernel module (something concerning xino handling, see the commit for more details), and so I had to introduce a lock around doing a mount() syscall so no two mounts happen concurrently.