@kolyshkin 2019-05-10
The error from `auplink flush` during benchmark is
> auplink:proc_mnt.c:96: /tmp/aufs-tests/aufs/mnt/36f594c070a6576ebc1a952da4c1d493067c7b134a979d636dfb6fe34e65993c: Invalid argument

comes from https://sourceforge.net/p/aufs/aufs-util/ci/aufs3.2/tree/proc_mnt.c#l96

and means that auplink can't find this mount in /proc/self/mounts.