@kolyshkin 2019-05-10
> and means that auplink can't find this mount in /proc/self/mounts.

My theory is this happens because the kernel can not guarantee the continuity of /proc/self/mounts contents, i.e. when many mounts are added and removed, while someone is reading /proc/self/mounts, its contents might be broken. Let me try to prove it...